+++ components/editing/BranchesSelector.tsx
@@ -1,31 +1,22 @@
 import React, { Suspense } from 'react'
-import { getCookieSession } from '@/components/lib/session'
-import { fetchBranches } from '@/components/lib/git-functions'
-import BranchesSelectorInner from '@/components/BranchesSelectorInner'
+import BranchesSelectorInner from '@/components/editing/BranchesSelectorInner'
 import { RepositoryRefInfo } from '@/components/context/EditorProvider'
 import SelectMenuLoading from '@/components/SelectMenuLoading'
+import { Branch } from '@/lib/provider/provider'
 
 export interface BranchesSelectorProps {
   repositoryInfo: RepositoryRefInfo
+  branches: Promise<Branch[]>
 }
 
 const BranchesSelector: React.FC<BranchesSelectorProps> = (
   props: BranchesSelectorProps,
 ) => {
-  const getBranches = async () => {
-    const session = await getCookieSession()
-    return fetchBranches(
-      session,
-      props.repositoryInfo.owner,
-      props.repositoryInfo.repository,
-    )
-  }
-
   return (
     <Suspense fallback={<SelectMenuLoading />}>
       <BranchesSelectorInner
         repositoryInfo={props.repositoryInfo}
-        branches={getBranches()}
+        branches={props.branches}
       />
     </Suspense>
   )